refactor: extracted firebase logic to custom hook
