+++ src/modules/PlanTripModule.tsx
@@ -2,71 +2,12 @@ import { Navigation } from 'react-feather'
 import { BaseIconInput } from '../components/atoms'
 import { MapyczMap } from '../components/molecules'
 import { END_AT_INPUT_ID, START_FROM_INPUT_ID } from '../const'
-import { setDoc, onSnapshot, updateDoc, deleteDoc } from 'firebase/firestore'
-import {
-  RecentlySearchedTrips,
-  recentlySearchedTripsCollection,
-  recentlySearchedTripsDocument,
-  recentlySearchedTripsDocumentById
-} from '../utils/firebase'
-import { useEffect, useState } from 'react'
-import { Coord } from '../types'
+import { usePlanTripFirebase } from '../hooks/planTrip'
 
 type LabeledInputProps = {
   id: string
   label: string
   placeholder: string
-}
-
-const createRecentlySearched = async (recentlySearchedTrips: RecentlySearchedTrips[], from: Coord, to: Coord) => {
-  /* TODO: filter also by userId (also in other 2 methods) */
-  var recentlySearchedByFromAndTo = recentlySearchedTrips.filter(
-    (x) => x.from.lat === from.lat && x.from.lng === from.lng && x.to.lat === to.lat && x.to.lng === to.lng
-  )
-
-  if (recentlySearchedByFromAndTo.length === 0) {
-    const newDocRef = recentlySearchedTripsDocument
-
-    await setDoc(newDocRef, {
-      id: newDocRef.id,
-      from: from,
-      to: to,
-      searchedOn: new Date(),
-      pinned: false,
-      /* TODO: set logged in user id */
-      userId: 'temp'
-    })
-  } else {
-    const idToBeUpdated = recentlySearchedByFromAndTo[0].id
-
-    const documentRef = recentlySearchedTripsDocumentById(idToBeUpdated)
-    await updateDoc(documentRef, {
-      searchedOn: new Date()
-    })
-  }
-}
-
-const updatePin = async (recentlySearchedTrips: RecentlySearchedTrips[], from: Coord, to: Coord) => {
-  var recentlySearchedByFromAndTo = recentlySearchedTrips.filter(
-    (x) => x.from.lat === from.lat && x.from.lng === from.lng && x.to.lat === to.lat && x.to.lng === to.lng
-  )
-
-  const idToBeUpdated = recentlySearchedByFromAndTo[0].id
-  const pinned = recentlySearchedByFromAndTo[0].pinned
-
-  const documentRef = recentlySearchedTripsDocumentById(idToBeUpdated)
-  await updateDoc(documentRef, {
-    pinned: !pinned
-  })
-}
-
-const deleteRecentlySearchedTrip = async (recentlySearchedTrips: RecentlySearchedTrips[], from: Coord, to: Coord) => {
-  var recentlySearchedByFromAndTo = recentlySearchedTrips.filter(
-    (x) => x.from.lat === from.lat && x.from.lng === from.lng && x.to.lat === to.lat && x.to.lng === to.lng
-  )
-
-  const idToBeUpdated = recentlySearchedByFromAndTo[0].id
-  await deleteDoc(recentlySearchedTripsDocumentById(idToBeUpdated))
 }
 
 const LabeledInput: React.FC<LabeledInputProps> = ({ label, ...inputProps }) => {
@@ -81,17 +22,7 @@ const LabeledInput: React.FC<LabeledInputProps> = ({ label, ...inputProps }) => 
 }
 
 const PlanTripModule: React.FC = () => {
-  const [recentlySearchedTrips, setrecentlySearchedTrips] = useState<RecentlySearchedTrips[]>([])
-
-  useEffect(() => {
-    const unsubscribe = onSnapshot(recentlySearchedTripsCollection, (snapshot) => {
-      setrecentlySearchedTrips(snapshot.docs.map((doc) => doc.data()))
-    })
-
-    return () => {
-      unsubscribe()
-    }
-  }, [])
+  const { recentlySearchedTrips, createRecentlySearched, deleteRecentlySearchedTrip, updatePin } = usePlanTripFirebase()
 
   return (
     <div className="flex -mt-10 3xl:container 3xl:mx-auto w-full grow">
